feat(jetstream): object store uses an interface
